adds pseudo code for more general email filter and test cases
