+++ src/peoplefinder/templatetags/email_word_break.py
@@ -2,13 +2,25 @@ from django import template
 from django.utils.html import escape
 from django.utils.safestring import mark_safe
 
-
 register = template.Library()
+
+
+
+# Split on @ -> name_part + domain_part
+
+# Name: split on .-_ -> name_parts ["name", "middle", "-name" ".surname"]
+# Domain: split on .-_ -> domain_parts ["example", ".com"]
+
+# For each name_part in name_parts: Split if longer than X
+# For each domain_part in name_parts: Split if longer than X
+
+# Add all parts back together with <wbr> being the glue
 
 
 @register.filter
 def email_word_break(email: str) -> str:
     if not email or not isinstance(email, str):
+        # raise typerror/valueerror, check python
         return email
 
     email = escape(email)
